builtin: Uncompress the builtin fonts
It's fairly stupid to compress these: you save a whopping 24k on disk,
but they're still going to be uncompressed in memory, only now those are
anonymous dirty pages that have to be preserved instead of a read-only
mmap that you could just evict.

The data here is copied from XFree86 3.9.16Z:

http://cvsweb.xfree86.org/cvsweb/*checkout*/xc/lib/font/builtins/fonts.c?rev=1.2

Also remove the call to BufFilePushZIP since it will now always fail.

Signed-off-by: Adam Jackson <ajax@redhat.com>