freetype: Fix a thinko
The code as written would build strings of the form "1-ISO10646" and
then act all surprised that they strcasecmp unequal to iso10646-1.

XXX this gets xlsfonts against the freetype renderer to not crash; it
still doesn't work right, all I can display is 'fixed' which is still
through the builtin path. But it's a start.

Signed-off-by: Adam Jackson <ajax@redhat.com>